Extract approvals helpers to shrink accounting-screen root
accounting-screen.tsx (7531 lines) held a self-contained cluster of 14 pure
approvals-queue helpers (status labels/tones, evidence summaries, blocker and
next-action derivation, small formatters) with no JSX and no back-references
into the root. Move them to accounting-screen.approvals.ts; 13 are re-imported
by the root, one stays internal. Root drops to 7372 lines. Same downward-closed
extraction pattern as the reporting report-writer helpers.
